refactor: separate to main_transform and run_transform
