+++ prestring/text/transform.py
@@ -38,7 +38,7 @@ def transform_file(fname: str, *, indent: str, m=None):
 
 if __name__ == "__main__":
     import sys
-    from prestring._cli import main_transform
+    from prestring.cli import main_transform
     from prestring.python import Module as PyModule
 
     main_transform(
